+++ fltk/src/prelude.rs
@@ -525,8 +525,8 @@ pub unsafe trait MenuExt: WidgetExt {
     /// Sets the text color
     fn set_text_color(&mut self, c: Color);
     /// Add a menu item along with its callback
-    /// The characters "&", "/", "\\", and "_" are treated as special characters in the label string. The "&" character specifies that the following character is an accelerator and will be underlined.
-    /// The "\\" character is used to escape the next character in the string. Labels starting with the "_" character cause a divider to be placed after that menu item.
+    /// The characters "&", "/", "\\", and "\_" (underscore) are treated as special characters in the label string. The "&" character specifies that the following character is an accelerator and will be underlined.
+    /// The "\\" character is used to escape the next character in the string. Labels starting with the "\_" (underscore) character cause a divider to be placed after that menu item.
     fn add<F: FnMut() + 'static>(
         &mut self,
         label: &str,
@@ -536,8 +536,8 @@ pub unsafe trait MenuExt: WidgetExt {
     ) where
         Self: Sized;
     /// Add a menu item along with its callback
-    /// The characters "&", "/", "\\", and "_" are treated as special characters in the label string. The "&" character specifies that the following character is an accelerator and will be underlined.
-    /// The "\\" character is used to escape the next character in the string. Labels starting with the "_" character cause a divider to be placed after that menu item.
+    /// The characters "&", "/", "\\", and "\_" (underscore) are treated as special characters in the label string. The "&" character specifies that the following character is an accelerator and will be underlined.
+    /// The "\\" character is used to escape the next character in the string. Labels starting with the "\_" (underscore) character cause a divider to be placed after that menu item.
     /// Takes the menu item as a closure argument
     fn add2<F: FnMut(&mut Self) + 'static>(
         &mut self,
@@ -548,8 +548,8 @@ pub unsafe trait MenuExt: WidgetExt {
     ) where
         Self: Sized;
     /// Inserts a menu item at an index along with its callback
-    /// The characters "&", "/", "\\", and "_" are treated as special characters in the label string. The "&" character specifies that the following character is an accelerator and will be underlined.
-    /// The "\\" character is used to escape the next character in the string. Labels starting with the "_" character cause a divider to be placed after that menu item.
+    /// The characters "&", "/", "\\", and "\_" (underscore) are treated as special characters in the label string. The "&" character specifies that the following character is an accelerator and will be underlined.
+    /// The "\\" character is used to escape the next character in the string. Labels starting with the "\_" (underscore) character cause a divider to be placed after that menu item.
     fn insert<F: FnMut() + 'static>(
         &mut self,
         idx: u32,
@@ -560,8 +560,8 @@ pub unsafe trait MenuExt: WidgetExt {
     ) where
         Self: Sized;
     /// Inserts a menu item at an index along with its callback
-    /// The characters "&", "/", "\\", and "_" are treated as special characters in the label string. The "&" character specifies that the following character is an accelerator and will be underlined.
-    /// The "\\" character is used to escape the next character in the string. Labels starting with the "_" character cause a divider to be placed after that menu item.
+    /// The characters "&", "/", "\\", and "\_" (underscore) are treated as special characters in the label string. The "&" character specifies that the following character is an accelerator and will be underlined.
+    /// The "\\" character is used to escape the next character in the string. Labels starting with the "\_" (underscore) character cause a divider to be placed after that menu item.
     /// Takes the menu item as a closure argument
     fn insert2<F: FnMut(&mut Self) + 'static>(
         &mut self,
@@ -573,8 +573,8 @@ pub unsafe trait MenuExt: WidgetExt {
     ) where
         Self: Sized;
     /// Add a menu item along with an emit (sender and message)
-    /// The characters "&", "/", "\\", and "_" are treated as special characters in the label string. The "&" character specifies that the following character is an accelerator and will be underlined.
-    /// The "\\" character is used to escape the next character in the string. Labels starting with the "_" character cause a divider to be placed after that menu item.
+    /// The characters "&", "/", "\\", and "\_" (underscore) are treated as special characters in the label string. The "&" character specifies that the following character is an accelerator and will be underlined.
+    /// The "\\" character is used to escape the next character in the string. Labels starting with the "\_" (underscore) character cause a divider to be placed after that menu item.
     fn add_emit<T: 'static + Clone + Send + Sync>(
         &mut self,
         label: &str,
@@ -585,8 +585,8 @@ pub unsafe trait MenuExt: WidgetExt {
     ) where
         Self: Sized;
     /// Inserts a menu item along with an emit (sender and message)
-    /// The characters "&", "/", "\\", and "_" are treated as special characters in the label string. The "&" character specifies that the following character is an accelerator and will be underlined.
-    /// The "\\" character is used to escape the next character in the string. Labels starting with the "_" character cause a divider to be placed after that menu item.
+    /// The characters "&", "/", "\\", and "\_" (underscore) are treated as special characters in the label string. The "&" character specifies that the following character is an accelerator and will be underlined.
+    /// The "\\" character is used to escape the next character in the string. Labels starting with the "\_" (underscore) character cause a divider to be placed after that menu item.
     fn insert_emit<T: 'static + Clone + Send + Sync>(
         &mut self,
         idx: u32,
@@ -600,8 +600,8 @@ pub unsafe trait MenuExt: WidgetExt {
     /// Remove a menu item by index
     fn remove(&mut self, idx: u32);
     /// Adds a simple text option to the Choice and MenuButton widgets
-    /// The characters "&", "/", "\\", "|", and "_" are treated as special characters in the label string. The "&" character specifies that the following character is an accelerator and will be underlined.
-    /// The "\\" character is used to escape the next character in the string. Labels starting with the "_" character cause a divider to be placed after that menu item.
+    /// The characters "&", "/", "\\", "|", and "\_" (underscore) are treated as special characters in the label string. The "&" character specifies that the following character is an accelerator and will be underlined.
+    /// The "\\" character is used to escape the next character in the string. Labels starting with the "\_" (underscore) character cause a divider to be placed after that menu item.
     fn add_choice(&mut self, text: &str);
     /// Gets the user choice from the Choice and MenuButton widgets
     fn choice(&self) -> Option<String>;