+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.14.9] - 2021-02-14
+### Changes
+- Add draw::draw_image().
+- Add OverlayWindow widget.
+- Relax requirement of static str for get_proc_address.
+- Update deps.
+
 ## [0.14.8] - 2021-02-13
 ### Changes
 - Add TabelExt::get_selection2 and visible_cells2 variants which return tuples.